pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Aleksandr Kirillov <tegos_9898@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.1.0"
Django = "^4.0.6"
django-bootstrap4 = "^22.1"
bootstrap4 = "^0.1.0"
flake8 = "^5.0.3"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
